ディレクトリ構成をMVCに変更
