--- app/pages/auth/index.ts
@@ -1,5 +1,0 @@
-import LoginPage from './LoginPage.js'
-
-export {
-  LoginPage,
-}
