--- app/servers/auth/index.ts
@@ -1,11 +1,0 @@
-import loginPassword from './loginPassword.js'
-import loginGoogleRedirect from './loginGoogleRedirect.js'
-import loginGoogleCallback from './loginGoogleCallback.js'
-import logout from './logout.js'
-
-export {
-  loginPassword,
-  loginGoogleRedirect,
-  loginGoogleCallback,
-  logout,
-}
