--- app/servers/auth/loginGoogleCallback.ts
@@ -1,12 +1,0 @@
-import type { Context } from 'hono'
-
-/**
- * Google認証のコールバック
- * @param c 
- * @returns 
- */
-const loginGoogleCallback = (c: Context) => {
-  return c.text('google callback')
-}
-
-export default loginGoogleCallback
