--- app/servers/auth/loginGoogleRedirect.ts
@@ -1,12 +1,0 @@
-import type { Context } from 'hono'
-
-/**
- * Google認証ページにリダイレクト
- * @param c 
- * @returns 
- */
-const loginGoogleRedirect = (c: Context) => {
-  return c.text('google redirect')
-}
-
-export default loginGoogleRedirect
