+++ core/adapters/handle.d.ts
