+++ core/adapters/index.ts
