+++ core/messages/locales/en.ts
@@ -12,6 +12,7 @@ const enTranslations = {
   loginAttemptExceeded: 'Login attempt exceeded',
   loginFailed: 'Login failed',
   loginSuccess: 'Login success',
+  logout: 'Logout',
 }
 
 export default enTranslations
