+++ core/messages/locales/ja.ts
@@ -12,6 +12,7 @@ const jaTranslations = {
   loginAttemptExceeded: 'ログイン試行回数が上限に達しました',
   loginFailed: 'ログインに失敗しました',
   loginSuccess: 'ログインに成功しました',
+  logout: 'ログアウトしました',
 }
 
 export default jaTranslations
