+++ tsconfig.json
@@ -10,8 +10,7 @@
     ],
     "jsx": "react-jsx",
     "jsxImportSource": "hono/jsx",
-    "outDir": "./dist",
-    "rootDir": "./app"
+    "outDir": "./dist"
   },
   "exclude": ["node_modules"]
 }
